refactror and centralize security logic
