+++ mobsf/DynamicAnalyzer/views/android/tests_frida.py
@@ -15,9 +15,9 @@ from django.views.decorators.http import require_http_methods
 from mobsf.DynamicAnalyzer.views.android.frida_core import Frida
 from mobsf.DynamicAnalyzer.views.common.shared import (
     invalid_params,
-    is_attack_pattern,
     send_response,
 )
+from mobsf.MobSF.security import is_attack_pattern
 from mobsf.DynamicAnalyzer.views.android.operations import (
     get_package_name,
 )
