+++ mobsf/DynamicAnalyzer/views/common/frida/views.py
@@ -9,9 +9,9 @@ from django.views.decorators.http import require_http_methods
 from mobsf.DynamicAnalyzer.views.common.shared import (
     send_response,
 )
+from mobsf.MobSF.security import is_safe_path
 from mobsf.MobSF.utils import (
     is_md5,
-    is_safe_path,
     print_n_send_error_response,
 )
 from mobsf.MobSF.views.authentication import (
