+++ mobsf/MobSF/utils.py
@@ -19,7 +19,6 @@ import stat
 import sqlite3
 import unicodedata
 import threading
-from urllib.parse import unquote
 from pathlib import Path
 from concurrent.futures import (
     ThreadPoolExecutor,
@@ -572,11 +571,6 @@ def read_sqlite(sqlite_file):
     return table_dict
 
 
-def is_pipe_or_link(path):
-    """Check for named pipe."""
-    return os.path.islink(path) or stat.S_ISFIFO(os.stat(path).st_mode)
-
-
 def get_network():
     """Get Network IPs."""
     ips = []
@@ -611,15 +605,6 @@ def get_proxy_ip(identifier):
     return proxy_ip
 
 
-def is_safe_path(safe_root, check_path, raw_file):
-    """Detect Path Traversal."""
-    if is_path_traversal(raw_file):
-        return False
-    safe_root = os.path.realpath(os.path.normpath(safe_root))
-    check_path = os.path.realpath(os.path.normpath(check_path))
-    return check_path.startswith(safe_root + os.sep) or check_path == safe_root
-
-
 def file_size(app_path):
     """Return the size of the file."""
     return round(float(os.path.getsize(app_path)) / (1024 * 1024), 2)
@@ -643,30 +628,6 @@ def get_config_loc():
         )
     else:
         return 'MobSF/settings.py'
-
-
-def clean_filename(filename, replace=' '):
-    if platform.system() == 'Windows':
-        whitelist = f'-_.() {string.ascii_letters}{string.digits}'
-        # replace spaces
-        for r in replace:
-            filename = filename.replace(r, '_')
-        # keep only valid ascii chars
-        cleaned_filename = unicodedata.normalize(
-            'NFKD', filename).encode('ASCII', 'ignore').decode()
-        # keep only whitelisted chars
-        return ''.join(c for c in cleaned_filename if c in whitelist)
-    return filename
-
-
-def cmd_injection_check(data):
-    """OS Cmd Injection from Commix."""
-    breakers = [
-        ';', '%3B', '&', '%26', '&&',
-        '%26%26', '|', '%7C', '||',
-        '%7C%7C', '%0a', '%0d%0a',
-    ]
-    return any(i in data for i in breakers)
 
 
 def strict_package_check(user_input):
@@ -711,45 +672,6 @@ def common_check(instance_id):
             'message': 'Invalid instance identifier'}
     else:
         return None
-
-
-def is_path_traversal(user_input):
-    """Check for path traversal."""
-    if not user_input:
-        return False
-
-    # Disallow absolute paths and windows paths and backslashes
-    if os.path.isabs(user_input) or user_input.startswith(('\\', '//')):
-        logger.error('Path traversal attack detected with absolute path')
-        return True
-
-    # Normalize and decode URL-encoded characters
-    try:
-        # Handle URL decoding (e.g., %2e -> .)
-        decoded = unquote(user_input)
-        # Handle double URL decoding (e.g., %252e -> %2e -> .)
-        double_decoded = unquote(decoded)
-    except Exception:
-        logger.error('Path traversal attack detected with invalid URL encoding')
-        return True
-
-    # Check for path traversal in both original and decoded versions
-    dangerous_patterns = ['..', '../', '..\\', '..\\\\']
-
-    # Check original filename
-    if any(pattern in user_input for pattern in dangerous_patterns):
-        logger.error('Path traversal attack detected with invalid path')
-        return True
-
-    # Check decoded versions
-    if any(pattern in decoded for pattern in dangerous_patterns):
-        logger.error('Path traversal attack detected with invalid path')
-        return True
-
-    if any(pattern in double_decoded for pattern in dangerous_patterns):
-        logger.error('Path traversal attack detected with invalid path')
-        return True
-    return False
 
 
 def is_zip_magic(file_obj):
