+++ mobsf/StaticAnalyzer/views/common/shared_func.py
@@ -23,14 +23,16 @@ from mobsf.MobSF import settings
 from mobsf.MobSF.security import (
     sanitize_for_logging,
 )
+from mobsf.MobSF.security import (
+    is_path_traversal,
+    is_safe_path,
+)
 from mobsf.MobSF.utils import (
     EMAIL_REGEX,
     STRINGS_REGEX,
     URL_REGEX,
     append_scan_status,
     is_md5,
-    is_path_traversal,
-    is_safe_path,
     print_n_send_error_response,
     set_permissions,
 )
